+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.10.4] - 22-SEPTEMBER-6
+
+- add `AxelarTransferAPI` to allow query transfer status for cross-chain transfer via deposit address or sendToken method.
+- rename `axelarCachingServiceUrl` to `axelarGMPApiUrl`
+
 ## [0.10.3] - 2022-SEPTEMBER-5
 
 - update the `AxelarGMPRecoveryAPI` to allow for rpc/lcd endpoint overrides in constructor for the API